take logic for base asset jobs out of AssetGroup
