+++ python_modules/dagster/dagster/_core/definitions/assets_job.py
@@ -1,3 +1,4 @@
+import warnings
 from collections import defaultdict
 from typing import Any, Dict, Iterable, List, Mapping, Optional, Sequence, Set, Tuple, Union
 
@@ -5,10 +6,10 @@ from toposort import CircularDependencyError, toposort
 
 import dagster._check as check
 from dagster._annotations import experimental
-from dagster._core.definitions.utils import DEFAULT_IO_MANAGER_KEY
 from dagster._core.errors import DagsterInvalidDefinitionError
 from dagster._core.selector.subset_selector import AssetSelectionData
 from dagster._utils import merge_dicts
+from dagster._utils.backcompat import ExperimentalWarning
 
 from .asset_layer import AssetLayer
 from .assets import AssetsDefinition
@@ -24,6 +25,60 @@ from .resource_definition import ResourceDefinition
 from .resource_requirement import ensure_requirements_satisfied
 from .source_asset import SourceAsset
 from .utils import DEFAULT_IO_MANAGER_KEY
+
+# Prefix for auto created jobs that are used to materialize assets
+ASSET_BASE_JOB_PREFIX = "__ASSET_JOB"
+
+
+def is_base_asset_job_name(name) -> bool:
+    return name.startswith(ASSET_BASE_JOB_PREFIX)
+
+
+def get_base_asset_jobs(
+    assets: Sequence[AssetsDefinition],
+    source_assets: Sequence[SourceAsset],
+    resource_defs: Optional[Mapping[str, ResourceDefinition]],
+    executor_def: Optional[ExecutorDefinition],
+) -> Sequence[JobDefinition]:
+    with warnings.catch_warnings():
+        warnings.simplefilter("ignore", category=ExperimentalWarning)
+
+        assets_by_partitions_def: Dict[
+            Optional[PartitionsDefinition], List[AssetsDefinition]
+        ] = defaultdict(list)
+        for assets_def in assets:
+            assets_by_partitions_def[assets_def.partitions_def].append(assets_def)
+
+        if len(assets_by_partitions_def.keys()) == 0 or assets_by_partitions_def.keys() == {None}:
+            return [
+                build_assets_job(
+                    name=ASSET_BASE_JOB_PREFIX,
+                    assets=assets,
+                    source_assets=source_assets,
+                    executor_def=executor_def,
+                    resource_defs=resource_defs,
+                )
+            ]
+        else:
+            unpartitioned_assets = assets_by_partitions_def.get(None, [])
+            jobs = []
+
+            # sort to ensure some stability in the ordering
+            for i, (partitions_def, assets_with_partitions) in enumerate(
+                sorted(assets_by_partitions_def.items(), key=lambda item: repr(item[0]))
+            ):
+                if partitions_def is not None:
+                    jobs.append(
+                        build_assets_job(
+                            f"{ASSET_BASE_JOB_PREFIX}_{i}",
+                            assets=assets_with_partitions + unpartitioned_assets,
+                            source_assets=[*source_assets, *assets],
+                            resource_defs=resource_defs,
+                            executor_def=executor_def,
+                        )
+                    )
+
+            return jobs
 
 
 @experimental
